Rename harmonised specs data file
