+++ src/website/src/pages/ec/components/site-wide/site-footer/demo/index.js
@@ -1,6 +1,6 @@
 import dataCore from '@ecl/specs-component-site-footer/demo/data-core--ec';
 import dataStandardised from '@ecl/specs-component-site-footer/demo/data-standardised--ec';
-import dataHarmonised from '@ecl/specs-component-site-footer/demo/data-harmonised-group1--ec';
+import dataHarmonised from '@ecl/specs-component-site-footer/demo/data-harmonised--ec';
 import template from '@ecl/twig-component-site-footer/site-footer.html.twig';
 import { correctSvgPath } from '@ecl/website-utils';
